Conceptualization and UI
Added in a couple of UI to make the game look nice before we add in actually Gameplay mechanics. First thing we'll have to do is figure out how to calculate the scaling of each block in order to fit 10 horizontally. After which, we work on block interactions (drawing out our block path).