Added Selection Icon; Collection Chain Validation
We added a Selection Icon that rotates around a block that has been highlighted.  We've also added a SelectionHandler to know what selection icon should enable. 

We've also included a base for the code used to validate and invalidate a chain, so that the blocks can either be destoryed, or send to your opponent.

A couple of frame drops here and there, but always goes back to 60fps whenever it feels like it. Next, we'll include a LineVisualization, and add a feature into Block.cs to handle like a QuadLinkedList for pathfinding.