Collision Improvements
Collision detection is either called every 100th of a second, or when a TriggerEnter2D is called, reducing the amount of callbacks for collision.
If all objects are grounded, and validation is executed, an event will trigger checking if all blocks are grounded.
Also, included a method that allows us to snap to the grid, keeping the playing field leveled and uniformed.