Update 2021.0.15.0v: Improved Game Over Condition
A Game Over shows up on the screen whenever  a block is outside of the PlayingField view. No random Game Overs, since there's a delayed timer in IMass that whenever times out, the block will be "old" thus allowing us to check if where it landed was outside of the PlayingField view.