Better configured Post-Processing Effects
I played around with the post processing for the Game Scene, and it's looking a lot better than I thought it would. I also when and modified the Block Images so that the outline of it is more defined when important in the game, and it's Image component's color material is changed. 
I also shift each color's saturation value a little higher, and bumped up the contrast, giving a very desirable and undistracting look into the game itself.