Update 2022.0.1.0v
Added in new post processing system, as well as improved on the visual department (which is looking stunning).
We've also fixed some errors/bugs regarding restarting the stage again. Note to self, don't always using a Singleton<T> class unless you really need it, because further down the line, it's just going to cause a lot of problems for everyone, and I don't enjoy it that much.